// src/app/models/project-item.ts
export class ProjectItem {
    companyName: string;
    companyLogoUrl: string;
    companyTechnology: string;
    country: string;
    shipDate: string;
    assigned: {
        userName: string;
        userAvatarUrl: string;
        userPosition: string;
    };
    status: number;
    type: number;
}